refactor : 명함 상세 페이지 헤더 예외 처리 수정
