+++ src/features/card-detail/components/empty.tsx
